backend/requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
ultralytics
streamlit
requests
